+++ transrss_manager/views.py
@@ -28,6 +28,9 @@ def feed_list(request: HttpRequest):
         return redirect('home')
     
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            raise PermissionDenied
+
         form = FeedAddForm(request.POST)
         if form.is_valid():
             feed = FeedSource(**form.cleaned_data)
@@ -48,6 +51,9 @@ def feed_detail(request: HttpRequest, id: int):
         })
     
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            raise PermissionDenied
+
         form = FeedAddForm(request.POST)
         if form.is_valid():
             feed.title = form.cleaned_data['title']
@@ -64,6 +70,9 @@ def feed_delete(request: HttpRequest, id: int):
     feed = get_object_or_404(FeedSource, pk=id)
 
     if request.method == 'POST':
+        if not request.user.is_authenticated:
+            raise PermissionDenied
+
         feed.delete()
         return redirect('home')
 
@@ -77,6 +86,9 @@ def matcher_list(request: HttpRequest, feed_id: int):
         return redirect('home')
     
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            raise PermissionDenied
+
         form = MatcherAddForm(request.POST)
         if form.is_valid():
             matcher = FeedMatcher(source=feed, **form.cleaned_data)
@@ -90,6 +102,9 @@ def matcher_list(request: HttpRequest, feed_id: int):
 
 def matcher_delete(request: HttpRequest, feed_id: int, matcher_id: int):
     if request.method == 'POST':
+        if not request.user.is_authenticated:
+            raise PermissionDenied
+
         matcher = get_object_or_404(FeedMatcher, pk=matcher_id)
         matcher.delete()
         return redirect('feed_detail', id=feed_id)
@@ -105,6 +120,9 @@ def match_download(request: HttpRequest):
     request : HttpRequest
         Http request instance.
     """
+    if not request.user.is_authenticated:
+        raise PermissionDenied
+
     client = Client(**TRANSMISSION_CONFIG)
     feeds = FeedSource.objects.all()
     for feed in feeds:
@@ -125,6 +143,9 @@ def match_download(request: HttpRequest):
 
 
 def torrent_refresh(request: HttpRequest):
+    if not request.user.is_authenticated:
+        raise PermissionDenied
+        
     suber = Process(target=feed_load)
     suber.daemon = True
     suber.start()
